use Faraday and inclusion_connect as a model
